Support disabling the CuttlefishService BluetoothChecker
Even if BluetoothChecker detects a real problem with the Bluetooth
implementation on Cuttlefish, tying it to VIRTUAL_DEVICE_BOOT_COMPLETED
causes problems with canary boot validation tests in infra, because it
disproportionately impacts users and test plans that may not need
functioning Bluetooth. If a Bluetooth test is run subsequently and
fails, this can still be detected and triaged separately.

The BluetoothChecker remains a valid and important boot completion check
for Cuttlefish users. If the Bluetooth adapter is enabled by the overlay
or manually by the user and does not become available, it is broken, and
this breakage needs to be fixed. Therefore, for normal development, the
BluetoothChecker remains enabled by default.

The BluetoothChecker can be disabled with:
-extra_bootconfig_args="androidboot.cuttlefish_service_bluetooth_checker=false"

Bug: 395709071
Flag: EXEMPT bugfix
Cherrypick-From: https://googleplex-android-review.googlesource.com/q/commit:921885a4446917d3ba8a0c0a01ff167a96267d1d
Merged-In: I1472b22be816147f59422f5ceabdee32ca55045e
Change-Id: I1472b22be816147f59422f5ceabdee32ca55045e